consolidate append logic.
