+++ clients/redshift/redshift.go
@@ -41,15 +41,6 @@ func (s *Store) Label() constants.DestinationKind {
 	return constants.Redshift
 }
 
-func (s *Store) Merge(tableData *optimization.TableData) error {
-	return shared.Merge(s, tableData, s.config, types.MergeOpts{
-		UseMergeParts: true,
-		// We are adding SELECT DISTINCT here for the temporary table as an extra guardrail.
-		// Redshift does not enforce any row uniqueness and there could be potential LOAD errors which will cause duplicate rows to arise.
-		SubQueryDedupe: true,
-	})
-}
-
 const (
 	describeNameCol        = "column_name"
 	describeTypeCol        = "data_type"
